chore(notfound-exception): implemetn error template to display all error exception when in production. update js, css, bootstrap links in app.j
