+++ assets/bootstrap.js
@@ -1,5 +1,7 @@
 import { startStimulusApp } from '@symfony/stimulus-bundle';
 
 const app = startStimulusApp();
+import 'bootstrap/dist/css/bootstrap.min.css';
+import 'bootstrap';
 // register any custom, 3rd party controllers here
 // app.register('some_controller_name', SomeImportedController);
