+++ templates/bundles/TwigBundle/Exception/error.html.twig
@@ -1,15 +1,34 @@
 {# templates/bundles/TwigBundle/Exception/error404.html.twig #}
-{% extends 'base.html.twig' %}
-{% block body %}
-    <h1>
-        Page not found
-    </h1>
-    <p>
-        The requested page couldn't be located. Checkout for any URL
-        misspelling or
-        <a href="{{ path('homepage') }}">
-            return to the homepage
-        </a>
-        .
-    </p>
-{% endblock %}
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <title>
+            404 - Page introuvable
+        </title>
+        <!-- Bootstrap CDN -->
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
+    </head>
+    <body class="bg-light">
+        <div class="container text-center mt-5">
+            <h1 class="">
+                {# LOGO #}
+                <a class="navbar-brand fw-bold" href="{{ path('app_home') }}">
+                    ❄ SnowTricks
+                </a>
+            </h1>
+            <p class="lead mb-4">
+                🏂 Oups ! Cette figure n'existe pas... ou vous avez raté la réception.
+            </p>
+            <!-- Image en chemin ABSOLU -->
+            {#<img src="public/images/errorpage-bg_Snowboard_crash_1" alt="Snowboard crash" class="img-fluid mb-4" style="max-width: 300px;">#}
+            <p>
+                La page que vous cherchez est introuvable.
+            </p>
+            <!-- Lien simple -->
+            <a href="{{ path('app_home') }}" class="btn btn-primary mt-3">
+                Retour à l'accueil
+            </a>
+        </div>
+    </body>
+</html>
